"Parse list length as long instead of integer
To accommodate larger lists and prevent potential overflow errors, listLength variable in JavaSortsMain class has been changed from an integer to long. All the relevant method adaptations have also been implemented accordingly."

+++ src/main/java/eu/lilithmonodia/javasorts/JavaFXSortsMain.java
@@ -81,7 +81,7 @@ public class JavaFXSortsMain extends Application {
 
             long listLength;
             try {
-                listLength = Integer.parseInt(input);
+                listLength = Long.parseLong(input);
             } catch (NumberFormatException exception) {
                 outputLabel.setText("Invalid input!");
                 return;
